Fix Using Specific Times as CLI Args
Using hour/minute times like `12:35` was causing parsing errors since
the chrono strptime function needs to be able to parse an entire Tm
struct. We fix this by constructing a string with the full Tm
information by prepending with the current date and appending the number
of seconds to the argument string, then attempting to parse it.

We also provide a nicer error message when the passed time is in the
past.